src: make fs functions handle pre-epoch timestamps
A wrong type cast prevented timestamps before 1970-01-01 from working
with functions like `fs.stat()`.

Fixes: https://github.com/nodejs/node/issues/32369

+++ src/node_file-inl.h
@@ -85,8 +85,7 @@ void FillStatsArray(AliasedBufferBase<NativeT, V8T>* fields,
                    static_cast<NativeT>(stat))
 
 #define SET_FIELD_WITH_TIME_STAT(stat_offset, stat)                          \
-  /* NOLINTNEXTLINE(runtime/int) */                                          \
-  SET_FIELD_WITH_STAT(stat_offset, static_cast<unsigned long>(stat))
+  SET_FIELD_WITH_STAT(stat_offset, static_cast<double>(stat))
 
   SET_FIELD_WITH_STAT(kDev, s->st_dev);
   SET_FIELD_WITH_STAT(kMode, s->st_mode);
